pyproject.toml
[project]
name = "fastapi-e-commerce-backend"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "bcrypt>=5.0.0",
    "fastapi[standard]>=0.135.1",
    "httpx>=0.28.1",
    "psycopg2-binary>=2.9.11",
    "pwdlib[argon2,bcrypt]>=0.3.0",
    "pydantic[email]>=2.12.5",
    "pytest>=9.0.2",
    "python-dotenv>=1.2.2",
    "python-jose[cryptography]>=3.5.0",
    "sqlalchemy>=2.0.48",
]
